support links in text-based browsers
This change modifies the list template to add
an additional <a> to the article header.

The existing a tag uses the `entry-link` css class
to make the containing article cliable by forcing
`position: absolute` in combination with an empty
anchor element `<a class='entry-link'></a>`

This has the effect of making the entire article
clickable to navigate in a graphical browser however
it is not rendered in a text-based browser.

To correct this the title text in the article header
is wrapped in a second <a> tag. The `entry-link` css
class is not applied and default css is inherited.

Closes: #1716